# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Admission & Entry Test Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - in: query
          name: testId
          schema:
            type: integer
        - in: query
          name: subject
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Add a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/bulk-import:
    post:
      operationId: bulkImportQuestions
      tags: [questions]
      summary: Bulk import questions from Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkQuestionsInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Test Sessions
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List all results
      parameters:
        - in: query
          name: testId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pass, fail]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultListResponse"
  /results/merit-list:
    get:
      operationId: getMeritList
      tags: [results]
      summary: Get merit list sorted by score
      parameters:
        - in: query
          name: testId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Merit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeritEntry"
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get result by ID with answers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
  /results/by-student/{studentId}:
    get:
      operationId: getResultByStudent
      tags: [results]
      summary: Get result by student ID
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # Interviews
...
components:
  schemas:
...
    BulkImportResult:
      type: object
      required: [imported, failed, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: string

    # Verification
...
    Question:
      type: object
      required: [id, testId, questionText, optionA, optionB, optionC, optionD, correctOption, marks, subject]
      properties:
        id:
          type: integer
        testId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        marks:
          type: integer
        subject:
          type: string
        orderIndex:
          type: integer
    QuestionInput:
      type: object
      required: [testId, questionText, optionA, optionB, optionC, optionD, correctOption, marks, subject]
      properties:
        testId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        marks:
          type: integer
        subject:
          type: string
        orderIndex:
          type: integer
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        marks:
          type: integer
        subject:
          type: string
        orderIndex:
          type: integer
    QuestionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BulkQuestionsInput:
      type: object
      required: [testId, questions]
      properties:
        testId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"

    # Test Session
...
    Answer:
      type: object
      required: [id, sessionId, questionId, selectedOption]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        questionId:
          type: integer
        selectedOption:
          type: string
        isCorrect:
          type: boolean
        markedForReview:
          type: boolean
...
    Result:
      type: object
      required: [id, sessionId, studentId, testId, totalScore, totalMarks, passingMarks, percentage, status, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        studentId:
          type: integer
        testId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        totalScore:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        percentage:
          type: number
        status:
          type: string
          enum: [pass, fail]
        rank:
          type: ["integer", "null"]
        createdAt:
          type: string
    ResultDetail:
      allOf:
        - $ref: "#/components/schemas/Result"
        - type: object
          properties:
            answers:
              type: array
              items:
                $ref: "#/components/schemas/Answer"
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    ResultListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    MeritEntry:
      type: object
      required: [rank, studentId, studentName, rollNumber, totalScore, percentage, status]
      properties:
        rank:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        rollNumber:
          type: string
        totalScore:
          type: integer
        percentage:
          type: number
        status:
          type: string
          enum: [pass, fail]
        program:
          type: ["string", "null"]

    # Interview
...