# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionCorrectOption } from './questionCorrectOption';

export interface Question {
  id: number;
  testId: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctOption: QuestionCorrectOption;
  marks: number;
  subject: string;
  orderIndex?: number;
}