# lib/api-zod/src/generated/types/questionListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface QuestionListResponse {
  data: Question[];
  total: number;
  page: number;
  limit: number;
}