# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateCorrectOption } from './questionUpdateCorrectOption';

export interface QuestionUpdate {
  questionText?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctOption?: QuestionUpdateCorrectOption;
  marks?: number;
  subject?: string;
  orderIndex?: number;
}